Refactor GameMode into data.type package
