+++ src/main/java/org/spongepowered/api/data/manipulator/immutable/entity/ImmutableGameModeData.java
@@ -27,7 +27,7 @@ package org.spongepowered.api.data.manipulator.immutable.entity;
 import org.spongepowered.api.data.manipulator.immutable.ImmutableVariantData;
 import org.spongepowered.api.data.manipulator.mutable.entity.GameModeData;
 import org.spongepowered.api.entity.living.Human;
-import org.spongepowered.api.entity.living.player.gamemode.GameMode;
+import org.spongepowered.api.data.type.GameMode;
 
 /**
  * An {@link ImmutableVariantData} for the current {@link GameMode} of a
