+++ src/main/java/org/spongepowered/api/entity/living/player/tab/PlayerTabInfo.java
@@ -25,7 +25,7 @@
 package org.spongepowered.api.entity.living.player.tab;
 
 import org.spongepowered.api.GameProfile;
-import org.spongepowered.api.entity.living.player.gamemode.GameMode;
+import org.spongepowered.api.data.type.GameMode;
 import org.spongepowered.api.text.Text;
 
 /**
