+++ src/main/java/org/spongepowered/api/text/selector/ArgumentTypes.java
@@ -27,7 +27,7 @@ package org.spongepowered.api.text.selector;
 import com.flowpowered.math.vector.Vector3d;
 import com.flowpowered.math.vector.Vector3i;
 import org.spongepowered.api.entity.EntityType;
-import org.spongepowered.api.entity.living.player.gamemode.GameMode;
+import org.spongepowered.api.data.type.GameMode;
 import org.spongepowered.api.scoreboard.Team;
 
 import java.util.Collection;
