+++ src/main/java/org/spongepowered/api/world/WorldCreationSettings.java
@@ -26,7 +26,7 @@ package org.spongepowered.api.world;
 
 import org.spongepowered.api.data.DataContainer;
 import org.spongepowered.api.entity.Entity;
-import org.spongepowered.api.entity.living.player.gamemode.GameMode;
+import org.spongepowered.api.data.type.GameMode;
 import org.spongepowered.api.world.gen.WorldGeneratorModifier;
 
 import java.util.Collection;
